python: Tighten types of various DistroInfo methods
A common pattern in `DistroInfo` and its subclasses is to have a
`result` parameter that controls the return value.  If it's `"object"`,
then the return value is a `DistroRelease`; otherwise, the return value
is one of various kinds of strings.  The type annotations currently deal
with this by simply saying that the return type is `DistroRelease | str`
or `list[DistroRelease | str]`.

As a maintainer of code that uses this module, I've found this quite
inconvenient in practice when trying to get `mypy` to pass.  The methods
that return lists are particularly awkward because `list` is invariant,
but even the methods that return single objects tend to require annoying
casts.

This commit adds lots of overloads to inform the type checker of what's
going on.  In the case of the list methods, this is an incompatible
change because neither of `list[DistroRelease | str]` and
`list[DistroRelease] | list[str]` can be assigned to the other.
However, since I think anyone who was trying to use the type annotations
here must have been relying on some casts anyway, I hope this isn't too
much of an inconvenience.

Signed-off-by: Colin Watson <cjwatson@debian.org>